+++ R/ggbetweenstats.R
@@ -65,8 +65,8 @@
 #'   With Tukey's method, outliers are below (1st Quartile) or above (3rd
 #'   Quartile) `outlier.coef` times the Inter-Quartile Range (IQR) (Default:
 #'   `1.5`).
-#' @param mean.plotting Logical that decides whether mean is to be highlighted
-#'   and its value to be displayed (Default: `TRUE`).
+#' @param mean.plotting Logical that decides whether centrality tendency measure
+#'   is to be displayed as a point with a label (Default: `TRUE`).
 #' @param point.args A list of additional aesthetic arguments to be passed to
 #'   the `geom_point` displaying the raw data.
 #' @param violin.args A list of additional aesthetic arguments to be passed to
@@ -175,7 +175,7 @@ ggbetweenstats <- function(data,
                            tr = 0.1,
                            mean.plotting = TRUE,
                            mean.point.args = list(size = 5, color = "darkred"),
-                           mean.label.args = list(size = 3),
+                           mean.label.args = list(size = 3, nudge_x = 0.4, segment.linetype = 4),
                            notch = FALSE,
                            notchwidth = 0.5,
                            outlier.tagging = FALSE,